+++ tests/end_to_end/test_update_end_to_end.py
@@ -1,9 +1,6 @@
-import shutil
-from collections import namedtuple
 from unittest.mock import ANY, patch
 
 import funcy
-import git
 import pytest
 from cookiecutter.utils import work_in
 from git import GitCommandError
@@ -13,52 +10,7 @@ import ballet.templating
 import ballet.update
 from ballet.compat import safepath
 from ballet.project import DEFAULT_CONFIG_NAME
-from ballet.templating import render_project_template
 from ballet.update import DEFAULT_BRANCH, TEMPLATE_BRANCH
-from tests.util import tree
-
-
-@pytest.fixture
-def quickstart(tempdir):
-    """
-    $ cd tempdir
-    $ ballet quickstart
-    $ tree .
-    """
-    # cd tempdir
-    with work_in(safepath(tempdir)):
-
-        project_slug = 'foo'
-        extra_context = {
-            'project_slug': project_slug,
-        }
-
-        # ballet quickstart
-        render_project_template(no_input=True,
-                                extra_context=extra_context,
-                                output_dir=safepath(tempdir))
-
-        # tree .
-        tree(tempdir)
-
-        repo = git.Repo(safepath(tempdir.joinpath(project_slug)))
-
-        yield (
-            namedtuple('Quickstart', 'tempdir project_slug repo')
-            ._make((tempdir, project_slug, repo))
-        )
-
-
-@pytest.fixture
-def project_template_copy(tempdir):
-    old_path = ballet.templating._get_project_template_path()
-    new_path = tempdir.joinpath('templates', 'project_template')
-    shutil.copytree(old_path, safepath(new_path))
-
-    with patch('ballet.templating._get_project_template_path') as m:
-        m.return_value = str(new_path)
-        tree(new_path)
-        yield new_path
 
 
 # Utility methods -------------------------------------------------------------
